libsrpc v0.3.0

+++ bench/bench_common.h
@@ -39,16 +39,16 @@ static inline void bench_report(const char *name, uint64_t *samples, size_t n,
     long double sum = 0.0L;
     for (size_t i = 0; i < n; i++) sum += (long double)samples[i];
 
-    double mean = (double)(sum / (long double)n);
+    double mean = (double)sum / (double)n;
     double p50 = (double)samples[n / 2];
-    double p90 = (double)samples[(size_t)(n * 0.90)];
-    double p99 = (double)samples[(size_t)(n * 0.99)];
+    double p90 = (double)samples[(size_t)((double)n * 0.90)];
+    double p99 = (double)samples[(size_t)((double)n * 0.99)];
 
     printf("%-28s n=%-7zu min=%8.3f;  p50=%8.3f;  p90=%8.3f;  p99=%9.3f;  max=%10.3f;  "
            "mean=%8.3f; мкс",
            name, n,
-           samples[0] / 1000.0, p50 / 1000.0, p90 / 1000.0,
-           p99 / 1000.0, samples[n - 1] / 1000.0, mean / 1000.0);
+           (double)samples[0] / 1000.0, p50 / 1000.0, p90 / 1000.0,
+           p99 / 1000.0, (double)samples[n - 1] / 1000.0, mean / 1000.0);
 
     if (wall_ns > 0) {
         double rate = (double)n / ((double)wall_ns / 1e9);
